requirements.txt
python-telegram-bot==13.15
python-dotenv
gspread
oauth2client
pyTelegramBotAPI==4.12.0
